requirements.txt
# requirements.txt

# Data Manipulation and ETL
pandas>=2.1.0
pyspark>=3.4.1

# Database Connectivity
psycopg2-binary>=2.9.1
sqlalchemy>=1.4.0

# Data Creation
faker>=13.0.0

# Environment Variables Management
python-dotenv>=0.21.0

# Additional Utilities
numpy>=1.23.0

# Optional: Data Validation and Quality
great-expectations~=0.17.10
